feat(remap transform): add error output stream
The details of the UI here are a little iffy, but this seemed like the
smallest change that was backwards-compatible and didn't introduce
a third option about what to do with the event when something goes
wrong. I fully expect some discussion and resulting changes here, but it
seemed worthwhile to put this up as a starting point.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>